Adds scripts/msr_header_gen.sh
Script requires loading the stock msr kernel module and uses
Intel's rdmsr tool.  The script must be run as root.

The script generates a C header file named xxyy.h where xx
is the CPU family and yy is the CPU model (both in hexidecimal).

The header file contains several useful #defines for the msr
emulation shim.  It also contains arrays of tuples containing
information about whether a certain MSR is readable and, if so,
what the value on a particular machine at a particular moment
in time happened to be.

Potential issues:
        1. Very Intel-centric.  AMD may get a separate script.
        2. Write-only MSRs are incorrectly marked as invalid.
                (Not sure if any exist.)
        3. No information about bitwise permissions, or which
                of these MSRs are undocumented.  Given prior
                efforts, I doubt this information will ever
                be added.
        4. No indication of which MSRs should be monotonically
                increasing.